requirements.txt
# Core ML dependencies
numpy==1.19.5
scipy==1.7.3
opencv-python-headless==4.5.5.64
dlib==19.22.0
face-recognition-models==0.3.0
face-recognition==1.3.0
Pillow==8.4.0

# Web framework
flask==2.0.3
werkzeug==2.0.3
gunicorn==20.1.0
python-dotenv==0.19.2
click==8.0.4

# Data handling
pandas==1.3.5
pymongo==3.12.3
openpyxl==3.0.9

# Production dependencies
whitenoise==5.3.0
python-multipart==0.0.5
